场景另存为。

+++ ShadowEditor.Web/src/event/menu/scene/NewSceneEvent.js
@@ -13,22 +13,30 @@ NewSceneEvent.prototype = Object.create(MenuEvent.prototype);
 NewSceneEvent.prototype.constructor = NewSceneEvent;
 
 NewSceneEvent.prototype.start = function () {
-    this.app.on('mNewScene.' + this.id, this.onNewScene.bind(this));
+    this.app.on(`mNewScene.${this.id}`, this.onNewScene.bind(this));
 };
 
 NewSceneEvent.prototype.stop = function () {
-    this.app.on('mNewScene.' + this.id, null);
+    this.app.on(`mNewScene.${this.id}`, null);
 };
 
 NewSceneEvent.prototype.onNewScene = function () {
     var editor = this.app.editor;
 
-    // UI.confirm('询问', '所有未保存数据将丢失，确定吗？', function (event, btn) {
-    //     if (btn === 'ok') {
-    editor.clear();
-    //     }
-    // });
-    document.title = '未命名';
+    if (editor.sceneName == null) {
+        editor.clear();
+        document.title = '未命名';
+        return;
+    }
+
+    UI.confirm('询问', '所有未保存数据将丢失，确定要新建场景吗？', function (event, btn) {
+        if (btn === 'ok') {
+            editor.clear();
+            editor.sceneID = null;
+            editor.sceneName = null;
+            document.title = '未命名';
+        }
+    });
 };
 
 export default NewSceneEvent;
